moved train staff classes to the upper folder /java
