+++ src/main/java/TrainOperator.java
@@ -1,40 +1,17 @@
-import java.util.HashMap;
+import model.staff.Employee;
 
-public class TrainOperator extends Employee{
-
+public class TrainOperator extends Employee {
 
     public TrainOperator(String name, int num) {
         super(name, num);
 
     }
 
-    /**
-     * returns what this employee obejct is being paid
-     */
     @Override
     public double getMonthlySalary() {
-        return super.getBaseSalary();
+        return super.getMonthlySalary() + 1000; // TODO change or whatever
     }
-
-    /**
-     * receives payment for an employee object
-     */
-    @Override
-    public void setPaid(boolean isPaid) {
-        TrainOperator.isPaid = isPaid;
-    }
-
-    /**
-     * returns Line that the Employee is assigned to
-     */
-    public String checkLine(Employee employee) {
-        return
-    }
-
-
-
 }
-
