--- src/main/java/model/train/TrainOperator.java
@@ -1,21 +1,0 @@
-package model.train;
-
-import model.staff.Employee;
-
-public class TrainOperator extends Employee {
-
-    public TrainOperator(String name, int num) {
-        super(name, num);
-
-    }
-
-    @Override
-    public double getMonthlySalary() {
-        return super.getMonthlySalary() + 1000; // TODO change or whatever
-    }
-}
-
-
-
-
-
